Locked orientation.

+++ app/src/main/java/com/durov/solutions/manager/model/Factor.kt
@@ -2,5 +2,5 @@ package com.durov.solutions.manager.model
 
 data class Factor(
     val name: String,
-    val priority: Float
+    val priority: Int
 )
